requirements.txt
anthropic>=0.25.0
Pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
jinja2>=3.1.0
apscheduler>=3.10.0
python-multipart>=0.0.9
